requirements.txt
streamlit==1.45.1
anthropic==0.52.0
pypdf==5.4.0
python-docx==1.1.2
requests==2.32.3
beautifulsoup4==4.13.3
lxml==5.3.1
reportlab==4.4.0
streamlit-local-storage==0.0.25
